// src/iviche/notificationStorage/models/NotificationType.ts
export enum NotificationType {
  USER_ACCOUNT_UPDATE = 'USER_ACCOUNT_UPDATE',
  USER_ACCOUNT_STATUS_CHANGE = 'USER_ACCOUNT_STATUS_CHANGE',
  NEWS_AND_POLLS_DIGEST = 'NEWS_AND_POLLS_DIGEST',
  POLL_MODERATION = 'POLL_MODERATION',
  VOTED_POLL_STATUS_CHANGE = 'VOTED_POLL_STATUS_CHANGE',
  POLL_STATUS_CHANGE = 'POLL_STATUS_CHANGE',
  NEW_POLL = 'NEW_POLL',
}